+++ src/main/java/insane96mcp/survivalreimagined/SurvivalReimagined.java
@@ -70,6 +70,7 @@ public class SurvivalReimagined
         SRMobEffects.MOB_EFFECTS.register(modEventBus);
         SRBlocks.BLOCKS.register(modEventBus);
         SRItems.ITEMS.register(modEventBus);
+        SREnchantments.ENCHANTMENTS.register(modEventBus);
         SRGlobalLootModifiers.LOOT_MODIFIERS.register(modEventBus);
         SRLootItemConditions.LOOT_ITEM_CONDITION_TYPES.register(modEventBus);
         Weights.initMaterialWeight();