+++ src/main/java/insane96mcp/survivalreimagined/setup/SREnchantments.java
@@ -1,0 +1,14 @@
+package insane96mcp.survivalreimagined.setup;
+
+import insane96mcp.survivalreimagined.SurvivalReimagined;
+import insane96mcp.survivalreimagined.module.experience.enchantment.Magnetic;
+import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraftforge.registries.DeferredRegister;
+import net.minecraftforge.registries.ForgeRegistries;
+import net.minecraftforge.registries.RegistryObject;
+
+public class SREnchantments {
+    public static final DeferredRegister<Enchantment> ENCHANTMENTS = DeferredRegister.create(ForgeRegistries.ENCHANTMENTS, SurvivalReimagined.MOD_ID);
+
+    public static final RegistryObject<Enchantment> MAGNETIC = ENCHANTMENTS.register("magnetic", Magnetic::new);
+}